Refactoring and protection agains URL Tampering

+++ CarQuest.Web/Controllers/CarController.cs
@@ -1,7 +1,5 @@
 ﻿namespace CarQuest.Web.Controllers;
 
-using Data.Models;
-
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 
@@ -9,44 +7,74 @@ using Services.Interfaces;
 
 using ViewModels.Car;
 
+using Car = Data.Models.Car;
+
 [Authorize]
 public class CarController : BaseController
 {
 	private readonly ICarService carService;
+	private readonly IMechanicService mechanicService;
 
-	public CarController(ICarService carService)
+	public CarController(ICarService carService, IMechanicService mechanicService)
 	{
 		this.carService = carService;
+		this.mechanicService = mechanicService;
 	}
 
 	public async Task<IActionResult> All()
 	{
-		IEnumerable<Car> cars = await carService.AllUserCarsAsync(GetUserId());
+		Guid userId = GetUserId();
+
+		if (await mechanicService.MechanicExistsByUserIdAsync(userId))
+		{
+			return RedirectToAction("Index", "Home");
+		}
+
+		IEnumerable<Car> cars = await carService.AllUserCarsAsync(userId);
 
 		return View(cars);
 	}
 
 	[HttpGet]
-	public IActionResult Add()
+	public async Task<IActionResult> Add()
 	{
+		if (await mechanicService.MechanicExistsByUserIdAsync(GetUserId()))
+		{
+			return RedirectToAction("Index", "Home");
+		}
+
 		return View();
 	}
 
 	[HttpPost]
 	public async Task<IActionResult> Add(CarAddAndUpdateViewModel car)
 	{
-		await carService.AddUserCarAsync(GetUserId(), car);
+		Guid userId = GetUserId();
 
-		if (ModelState.IsValid)
+		if (await mechanicService.MechanicExistsByUserIdAsync(userId))
 		{
-			return RedirectToAction("All");
+			return RedirectToAction("Index", "Home");
 		}
 
-		return View();
+		if (!ModelState.IsValid)
+		{
+			return View();
+		}
+		
+		await carService.AddUserCarAsync(userId, car);
+
+		return RedirectToAction("All");
 	}
 
 	public async Task<IActionResult> Remove(Guid Id)
 	{
+		Guid userId = GetUserId();
+
+		if (await mechanicService.MechanicExistsByUserIdAsync(userId))
+		{
+			return RedirectToAction("Index", "Home");
+		}
+
 		await carService.DeleteUserCarAsync(Id);
 
 		return RedirectToAction("All");
@@ -55,6 +83,13 @@ public class CarController : BaseController
 	[HttpGet]
 	public async Task<IActionResult> Edit(Guid Id)
 	{
+		Guid userId = GetUserId();
+
+		if (await mechanicService.MechanicExistsByUserIdAsync(userId))
+		{
+			return RedirectToAction("Index", "Home");
+		}
+
 		CarAddAndUpdateViewModel car = await carService.GetCarAddAndUpdateViewModelAsync(Id);
 
 		return View(car);
@@ -63,6 +98,18 @@ public class CarController : BaseController
 	[HttpPost]
 	public async Task<IActionResult> Edit(Guid Id, CarAddAndUpdateViewModel car)
 	{
+		Guid userId = GetUserId();
+
+		if (await mechanicService.MechanicExistsByUserIdAsync(userId))
+		{
+			return RedirectToAction("Index", "Home");
+		}
+
+		if (!ModelState.IsValid)
+		{
+			return View(car);
+		}
+
 		await carService.UpdateUserCarAsync(Id, car);
 
 		return RedirectToAction("All");
